new assertNetworkConnection function - only switch network before transactions
