+++ src/pages/evm/wrap/WrapTab.vue
@@ -92,14 +92,7 @@ onBeforeMount(() => {
 
 async function handleCtaClick() {
     const label = CURRENT_CONTEXT;
-    if (!await accountStore.isConnectedToCorrectNetwork(label)) {
-        const networkName = useChainStore().loggedChain.settings.getDisplay();
-        const errorMessage = ant.config.localizationHandler('evm_wallet.incorrect_network', { networkName });
-
-        ant.config.notifyFailureWithAction(errorMessage, {
-            label: ant.config.localizationHandler('evm_wallet.switch'),
-        });
-
+    if (!await useAccountStore().assertNetworkConnection(label)) {
         return;
     }
